IconFlow/requirements.txt
pyqt5
pillow
# cairosvg is commented out due to missing cairo dependency on Windows
# cairosvg
